Move Position and MatrixProjection to a separate module
