+++ tests/test_connector.py
@@ -1,13 +1,4 @@
-import pytest
-
-from gaphas.connector import Handle, Position
-
-
-@pytest.mark.parametrize("position", [(0, 0), (1, 2)])
-def test_position(position):
-    pos = Position(position)
-    assert position[0] == pos.x
-    assert position[1] == pos.y
+from gaphas.connector import Handle
 
 
 def test_handle_x_y():
